Fixup SSH key provisioning service

+++ modules/profiles/server.nix
@@ -61,16 +61,14 @@
   systemd.services."default-ssh-keys" = {
     script = ''
       mkdir /root/.ssh
-      cat /boot/default-ssh-authorized-keys.txt >> /root/.ssh/authorized_keys
-    '';
-    wantedBy = [ "sshd.service" "sshd.socket" ];
-    unitConfig = {
-      ConditionPathExists = [ "!/root/.ssh/authorized_keys" "/boot/default-ssh-authorized-keys.txt" ];
-    };
-  };
 
-  systemd.services."generate-ssh-key" = {
-    script = ''
+      # Try to copy keys from the ESP
+      if [ -e /boot/default-ssh-authorized-keys.txt ]; then
+        cat /boot/default-ssh-authorized-keys.txt >> /root/.ssh/authorized_keys
+        exit 0
+      fi
+
+      # Otherwise, generate a default keypair and display it
       ${pkgs.openssh}/bin/ssh-keygen -f /root/.ssh/id_default -t ed25519
       cat /root/.ssh/id_default.pub > /root/.ssh/authorized_keys
       if [ -e /dev/ttyS0 ]; then
@@ -81,7 +79,6 @@
       fi
     '';
     wantedBy = [ "sshd.service" "sshd.socket" ];
-    conflicts = [ "default-ssh-keys.service" ];
     unitConfig = {
       ConditionPathExists = [ "!/root/.ssh/authorized_keys" ];
     };
